backend/requirements.txt
africastalking==1.2.8
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
